Extract sorting from WebSpeechVoiceManager
